pyproject.toml
[tool.poetry]
name = "dailybkup"
version = "0.1.0"
description = ""
authors = ["Vitor Barbosa <vitorqb@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.4.1"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
mypy = "^0.961"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
